pyproject.toml
[project]
name = "boot-dev-ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]

[tool.uv.workspace]
members = [
    "boot_dev_ai_project",
]
